+++ testing/test_examples.py
@@ -12,4 +12,4 @@ def test_telemetry_signalk_demo(caplog):
 
 def test_telemetry_nmea0183_demo(caplog):
     calypso_nmea0183_telemetry_demo(port=60110)
-    assert "Sending message to udp://255.255.255.255:60110. $IIVWR,154.0,L,11.06,N,5.69,M,20.48,K*65" in caplog.messages
+    assert "Sending message to udp://255.255.255.255:60110. $MLVWR,154.0,L,11.06,N,5.69,M,20.48,K*64" in caplog.messages